Added login app and modified login and register .m files
LoginRegister app was created to have a dedicated application for logging in and registering accounts, which then calls the Steg app and closes the LoginRegister app when a successful login was performed. The two .m files: logindlg and registerdlg were both edited to accommodate for the addition of the new app.